switch to LGPLv3
With the exception of code related to the point triangulation licensed under BSD2 from Anipose, I have written all of this code myself. Licensing had originally been selected with the intent of merging with another AGPL3 project, but it seems like that is unlikely now, so defaulting to a more permissive license.