update LICENSE
Pyxy3D uses triangulation code adapted from Anipose which is BSD 2-clause. The only other contributor to this project is rgov (Ryan Gosteves) who agreed in issue #600 with the switch to the BSD 2-clause.

I would like to make this code available with fewer strings attached, therefore I have made the decision to switch to the BSD 2-Clause. It should be noted that a primary dependency of Pyxy3D is PySide6 which is licensed under the LGPLv3. Pyxy3D does not make any modifications to the source code of PySide6, which is available via PyPI.