Diagnosing failed calibration
Using user provided calibration data discovered how things degrade when
there is an "unlinked" camera in the calibration. It is possible to
resolve some errors in intermediate table references and get the
calibration to run, but the results are clearly corrupted.

This provides a basis for building out more robust handling of unlinked
camera which may enable an expanded functionality in the future.