698 create pointestimates from subset of cameras (#709)
* Create a failing test

This test fails when attempting to optimize the camera array. The shape
of the point estimates currently does not align with the shape of the
camera array because there is one camera (5) which is unlinked with the
rest, so has no initial intrinsic values.

* remove old version of function no longer referenced

* corner_id -> point_id

the corner_id naming is overly specific, particularly as in the near
future I want to experiement with methods of initializing the capture
volume estimate that do not rely on calibration board or aruco markers.

* rename port_index and pose_index methods

These are now renamed to better reflect that it is only showing the
posed cameras. Trying to make the code better self documenting after
looking at things from the point of view of `get_point_estimates`.

* integrate new cam array methods into capture volume

The revised interface of the camera array was causing problems with how
the rmse summary was accessed.This cleans that up by changing how rmse
calls out to the camera array.

* new tests with unlinked passing

But optimization hangs because underlying data is bad

* handle unlinked cameras

Need to make sure that we are only using linked cameras in the point
estimates

* Allow flag to ignore camera for calibration

Cam 4 in current tests is very bad intrinsics. Want to be able to try to
calibrate by excluding a Camera without going and deleting all its data

* include ignore if it's there

use of this isn't currently being implemented anywhere, but this is a
first step.

* ignore cam 4 on test case

more efficient than deleting all cam 4 data

* adjust so now shouldn't have cams 4 or 5 (for different reasons)

* successfully running primary optimization

Apply fix to only reference posed cameras. Note that complete test is
not running successfully right now, there is a problem in the
QualityController where points are filtered which is causing a failure.

* correct comments

* track optimization loops better

* adjust for ignored camera

* comment out filtering; fixing this is newly added issue

Filtering fails for the calibration edge case here. Resolving that is
out of scope of the current branch.

* TODO associated with new issue 707

* breakpoint placeholder added to test

* ruff autoformatting