add support for fisheye lens in calibration
While the camera data object had previously been updated to allow
storage of a fisheye calibration and application of its methods, the
intrinsic calibrator wasn't adjusting the calibration method based on
the fisheye property. This commit resolves that. Note that this is not
tested on data with confirmed FOV > 150 degrees.