frame by frame assertions to test triangulator (#752)
* minor tweaks to reference for csv file

* enable easier charuco triangulation comparison

Useful for confirming that optimized point_estimates method and
triangulated 2d points with an optimized camera array produce 3d points
within tolerance.

* comparison oftne working, but missing points

while most of the points are within a mm, occasionaly there will be a
point that is 172mm off. not sure what is going on. Locking this down
before the next round of revisions.

* refine comparisons running on 3d points

clearer to see that the bundle adjustment and the triangulation are in
good agreement.