create aruco tracker
Creates a new Tracker that can integrate with the stream managers to generate 2D point files. There is an option to use mirror processing which could potentially be used for a large "calibration flag" that has a mirror image of the ArUco on the opposite side. 

This tracker wll be used as the basis of the next major milestone, which is the alternate calculation of the stereoocalibration.