771 make imagepoints the single source of truth for 2d observations (#777)
XYData is now ImagePoints and is created with ImagePoints.from_csv rather than passing a csv path to another function/method. Lots of cleanup in tests and controller to align with this. took the opportunity to clean up pytest.yml to allow controller testing on ubuntu (had been failing because pytest was running ubuntu headless).