paired pose network with legacy xy charuco (#803)
Picking up with development script for PnP workflow here. Previously we were matching the gold standard based on the old stereocalibration workflow which has now been phased out. This is a baby-step to store to the gold standard using the current framework. This lays the foundation for aligning with the pnp workflow with the new data structures. 

Essenitally, we are going to try to wrap up the experimental approach into a new version of estimate_paired_pose_network.