face keypoint depth map for markerless calibration
Uses an idealized 3D face mesh to provide keypoint x,y,x values in a single "face frame of reference". These 3d points slot into the calibration workflow that is now built off of pnp rather than stereocalibration.

A script to demonstrate this as a proof-of-concept is included, though in the process of doing this I have begun to get clarity on how to refine the structure of the code base for improved composability/flexibility to facilitate future developments. After this "new feature" I suspect i'm about to pivot back to tech-debt-repayment mode.