Create PointData Bundle Structure
The capture volume has not been phased out as the means of running the optimization, but this PR brings us closer to doing so. The data in the CaptureVolume is now more cleanly represented by the ImagePoints and WorldPoints in a PointDataBundle with the camera arrray. This pulls togetther everything needed for the optimization with more clean naming conventions. This also allows more clean handling of data persistence,